feat(director-ai): v0.4.0 — NLI scorer, vector store, streaming kernel, physics bridge, SSGF cycle, Lyapunov proofs
6 new modules, 7 new test files, 100 new tests (44→144), 86.93% coverage,
0 mypy errors (enforced), Sphinx docs, PyPI publish workflow.

New core modules: NLIScorer (DeBERTa + fallback), VectorGroundTruthStore
(InMemory/ChromaDB), StreamingKernel (3 halt mechanisms),
PhysicsBackedScorer (L16→consumer bridge).

New research modules: SSGFEngine (geometry learning with TCBO/PGBO),
LyapunovProof (symbolic SymPy + numerical Monte Carlo).

Cross-validation: 14 tests against SCPN-CODEBASE canonical parameters.

+++ src/director_ai/research/physics/sec_functional.py
@@ -124,7 +124,7 @@ class SECFunctional:
         entropy = -float(np.sum(p * np.log(p)))
         # Penalise low entropy (concentration)
         max_entropy = np.log(n_bins)
-        return self.lambda_entropy * (max_entropy - entropy)
+        return float(self.lambda_entropy * (max_entropy - entropy))
 
     def evaluate(
         self,
@@ -189,7 +189,7 @@ class SECFunctional:
             return 0.0
         # Approximate g(0) as 1 / (√(2π) · σ_ω)
         g0 = 1.0 / (np.sqrt(2 * np.pi) * std_omega)
-        return 2.0 / (np.pi * g0)
+        return float(2.0 / (np.pi * g0))
 
     def is_stable(self, sec_result: SECResult, tolerance: float = 1e-3) -> bool:
         """Check if dV/dt ≤ 0 (Lyapunov stability condition)."""
